requirements.txt
anthropic>=0.42.0
requests>=2.32.0
PyYAML>=6.0.2
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-cov>=6.0.0
ruff>=0.8.0
mypy>=1.13.0
